requirements.txt
Pillow
numpy
waitress
torch
requests
wtforms
opencv-python
werkzeug
flask
